currency_converter/pyproject.toml
[project]
name = "currency-converter"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "currency_app", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
